Asset management complete

+++ src/components/TransformComponent.h
@@ -6,10 +6,10 @@ struct TransformComponent
 {
     glm::vec2 m_position;
     glm::vec2 m_scale;
-    float m_rotation;
+    double m_rotation;
 
     TransformComponent() = default;
 
-    explicit TransformComponent(const glm::vec2 position, const glm::vec2 scale, const float rotation) : m_position(position), m_scale(scale), m_rotation(rotation)
+    explicit TransformComponent(const glm::vec2 position, const glm::vec2 scale, const double rotation) : m_position(position), m_scale(scale), m_rotation(rotation)
     {}
 };
